# tools/openapi-generator/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/exercises/workout/{id}": {
      "get": {
        "operationId": "ExerciseController_findByWorkout",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Exercise"
        ]
      }
    },
...
    "/workout-templates": {
      "post": {
        "operationId": "WorkoutTemplateController_createWorkoutTemplate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkoutTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "WorkoutTemplate"
        ]
      }
    },
    "/workout-templates/plan/{id}": {
      "get": {
        "operationId": "WorkoutTemplateController_getWorkoutTemplatesByWorkoutPlanId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "WorkoutTemplate"
        ]
      }
    },
    "/workout-templates/{id}": {
      "get": {
        "operationId": "WorkoutTemplateController_getWorkoutTemplateById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "WorkoutTemplate"
        ]
      }
    },
...
    "/workout-session/detail": {
      "get": {
        "operationId": "WorkoutSessionController_getWorkoutSessionDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "WorkoutSession"
        ]
      }
    },
    "/workout-session": {
      "post": {
        "operationId": "WorkoutSessionController_createWorkoutSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkoutSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "WorkoutSession"
        ]
      }
    },
    "/workout-session/{id}": {
      "delete": {
        "operationId": "WorkoutSessionController_deleteWorkoutSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "WorkoutSession"
        ]
      },
      "put": {
        "operationId": "WorkoutSessionController_updateWorkoutSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkoutSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "WorkoutSession"
        ]
      }
    },
    "/workout-session/{id}/exercise/{exerciseId}": {
      "delete": {
        "operationId": "WorkoutSessionController_deleteWorkoutSessionExercise",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exerciseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "WorkoutSession"
        ]
      }
    },
    "/workout-session/{id}/exercise/{exerciseId}/sets": {
      "put": {
        "operationId": "WorkoutSessionController_updateWorkoutSessionExercise",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exerciseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkoutSessionExerciseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "WorkoutSession"
        ]
      }
    },
    "/workout-session/plan/{id}": {
      "get": {
        "operationId": "WorkoutSessionController_getWorkoutSessionsByPlanId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "WorkoutSession"
        ]
      }
    },
    "/workout-session/muscle-group-stats": {
      "get": {
        "operationId": "WorkoutSessionController_getMuscleGroupStats",
        "parameters": [
          {
            "name": "periodType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "periodValue",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "WorkoutSession"
        ]
      }
    }
  },
  "info": {
    "title": "API",
    "description": "API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateWorkoutTemplateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "workoutPlanId": {
            "type": "string",
            "format": "uuid"
          },
          "exerciseIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "name",
          "workoutPlanId",
          "exerciseIds"
        ]
      },
...
      "CreateWorkoutSessionDto": {
        "type": "object",
        "properties": {
          "duration": {
            "type": "integer",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "completedAt": {
            "nullable": true
          },
          "workoutPlanId": {
            "type": "string"
          },
          "workoutTemplateId": {
            "type": "string"
          },
          "sessionExercises": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "exerciseId": {
                  "type": "string"
                },
                "exerciseName": {
                  "type": "string"
                },
                "notes": {
                  "type": "string"
                },
                "performedSets": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "reps": {
                        "type": "integer",
                        "nullable": true
                      },
                      "weight": {
                        "type": "number",
                        "nullable": true
                      },
                      "isCompleted": {
                        "type": "boolean"
                      },
                      "performedAt": {},
                      "setNumber": {
                        "type": "integer"
                      },
                      "restTime": {
                        "type": "integer",
                        "nullable": true
                      },
                      "distance": {
                        "type": "number",
                        "nullable": true
                      },
                      "duration": {
                        "type": "integer",
                        "nullable": true
                      }
                    },
                    "required": [
                      "id",
                      "reps",
                      "weight",
                      "isCompleted",
                      "performedAt",
                      "setNumber",
                      "restTime",
                      "distance",
                      "duration"
                    ]
                  }
                }
              },
              "required": [
                "exerciseId",
                "exerciseName",
                "performedSets"
              ]
            }
          }
        },
        "required": [
          "duration",
          "notes",
          "completedAt",
          "workoutPlanId",
          "workoutTemplateId",
          "sessionExercises"
        ]
      },
      "UpdateWorkoutSessionExerciseDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateWorkoutSessionDto": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "completedAt": {
            "type": "string"
          },
          "sessionExercises": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "order": {
                  "type": "integer"
                },
                "exerciseId": {
                  "type": "string"
                },
                "exerciseName": {
                  "type": "string"
                },
                "notes": {
                  "type": "string"
                },
                "performedSets": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "reps": {
                        "type": "integer",
                        "nullable": true
                      },
                      "weight": {
                        "type": "number",
                        "nullable": true
                      },
                      "isCompleted": {
                        "type": "boolean"
                      },
                      "setNumber": {
                        "type": "integer"
                      },
                      "restTime": {
                        "type": "integer",
                        "nullable": true
                      },
                      "distance": {
                        "type": "number",
                        "nullable": true
                      },
                      "duration": {
                        "type": "integer",
                        "nullable": true
                      }
                    },
                    "required": [
                      "id",
                      "reps",
                      "weight",
                      "isCompleted",
                      "setNumber",
                      "restTime",
                      "distance",
                      "duration"
                    ]
                  }
                }
              },
              "required": [
                "id",
                "order",
                "exerciseId",
                "exerciseName",
                "performedSets"
              ]
            }
          }
        }
      }
    }
  }
}